Split models.py into models package
Refs #34
