+++ resources/errors.py
@@ -1,0 +1,5 @@
+from django.core.exceptions import ValidationError
+
+
+class InvalidImage(ValidationError):
+    pass
