+++ resources/models/__init__.py
@@ -1,0 +1,4 @@
+from .availability import Day, Period, get_opening_hours  # noqa
+from .reservation import Reservation  # noqa
+from .resource import Purpose, Resource, ResourceType, ResourceImage  # noqa
+from .unit import Unit, UnitIdentifier  # noqa
