+++ resources/models/utils.py
@@ -1,0 +1,54 @@
+import datetime
+import time
+
+import base64
+import struct
+import arrow
+from django.conf import settings
+from django.utils import timezone
+
+DEFAULT_LANG = settings.LANGUAGES[0][0]
+
+def save_dt(obj, attr, dt, orig_tz="UTC"):
+    """
+    Sets given field in an object to a DateTime object with or without
+    a time zone converted into UTC time zone from given time zone
+
+    If there is no time zone on the given DateTime, orig_tz will be used
+    """
+    if dt.tzinfo:
+        arr = arrow.get(dt).to("UTC")
+    else:
+        arr = arrow.get(dt, orig_tz).to("UTC")
+    setattr(obj, attr, arr.datetime)
+
+
+def get_dt(obj, attr, tz):
+    return arrow.get(getattr(obj, attr)).to(tz).datetime
+
+
+def get_translated(obj, attr):
+    key = "%s_%s" % (attr, DEFAULT_LANG)
+    val = getattr(obj, key, None)
+    if not val:
+        val = getattr(obj, attr)
+    return val
+
+
+def generate_id():
+    t = time.time() * 1000000
+    b = base64.b32encode(struct.pack(">Q", int(t)).lstrip(b'\x00')).strip(b'=').lower()
+    return b.decode('utf8')
+
+
+def time_to_dtz(time, date=None, arr=None):
+    tz = timezone.get_current_timezone()
+    if time:
+        if date:
+            return tz.localize(datetime.datetime.combine(date, time))
+        elif arr:
+            return tz.localize(datetime.datetime(arr.year, arr.month, arr.day, time.hour, time.minute))
+    else:
+        return None
+
+
